split test into individual files
